스크립트 세팅 변경 적용, 기능추가 필요함

+++ Assets/Scripts/Unity/Object/MonsterObject.cs
@@ -27,7 +27,7 @@ namespace Client
                 transform.position = new Vector3(positionData.X, positionData.Y, positionData.Z);
 
 
-                int monsterHP = _monster.GetCurrentHP();
+                long monsterHP = _monster.GetCurrentHP();
                 if (monsterHP != currentHP)
                 {
                     currentHP = monsterHP;
